lots of tweaks
